+++ src/main/java/subway/views/InputView.java
@@ -1,9 +1,7 @@
 package subway.views;
 
 import subway.menus.LineMenu;
-import subway.menus.MainMenu;
 import subway.menus.SectionMenu;
-import subway.menus.StationMenu;
 
 import java.util.Scanner;
 
@@ -16,25 +14,7 @@ public class InputView {
         return scanner.nextLine();
     }
 
-    public static MainMenu selectMainMenu(Scanner scanner) {
-        try {
-            OutputView.printFeatureSelectMessage();
-            return MainMenu.getMenu(userInput(scanner));
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-            return selectMainMenu(scanner);
-        }
-    }
 
-    public static StationMenu selectStationMenu(Scanner scanner) {
-        try {
-            OutputView.printFeatureSelectMessage();
-            return StationMenu.getMenu(userInput(scanner));
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-            return selectStationMenu(scanner);
-        }
-    }
 
     public static LineMenu selectLineMenu(Scanner scanner) {
         try {
